requirement.txt
Django==3.1
Flask==1.1.2
h5py==2.8.0
image==1.5.32
imageio==2.9.0
imutils==0.5.3
Keras==2.3.1
matplotlib==3.1.3
mtcnn==0.1.0
mysql-connector-python==8.0.21
numpy==1.18.1
opencv-python==4.3.0.38
Pillow==7.2.0
protobuf==3.6.0
regex==2020.7.14
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.4.1
tensorflow==2.1.0
waitress==1.4.4
